Small changes related to cert generation
Add documentation to generate credentials using generate_cert.sh
Add check to see if pem files exist before removing them.
Use out parameter to avoid capturing stderr when running openssl
commands.
Add `set -x` to generate_cert.sh so it exits if any commands fail.

+++ proxy/certs/generate_cert.sh
@@ -1,3 +1,12 @@
-rm *.pem
-openssl genpkey -algorithm EC -pkeyopt ec_paramgen_curve:P-256 >& ca_private_key.pem
-openssl req -x509 -key ca_private_key.pem -subj /CN=proxy >& ca_cert.pem
+#!/bin/sh
+
+set -e
+
+CERTS_DIR="$(dirname -- "$(realpath -- "$0")")"
+
+if [[ -f *.pem ]]; then
+    rm *.pem
+fi
+
+openssl genpkey -algorithm EC -pkeyopt ec_paramgen_curve:P-256 -out ${CERTS_DIR}/ca_private_key.pem
+openssl req -x509 -key ${CERTS_DIR}/ca_private_key.pem -subj /CN=proxy -out ${CERTS_DIR}/ca_cert.pem
